It's follow-up patch for WL#7155.
This patch removes warning about comparison between signed and unsigned
integer expressions in method xid_t::get_my_xid().

+++ sql/xa.h
@@ -146,7 +146,8 @@ public:
 
   my_xid get_my_xid() const
   {
-    if (gtrid_length == MYSQL_XID_GTRID_LEN && bqual_length == 0 &&
+    if (gtrid_length == static_cast<long>(MYSQL_XID_GTRID_LEN) &&
+        bqual_length == 0 &&
         !memcmp(data, MYSQL_XID_PREFIX, MYSQL_XID_PREFIX_LEN))
     {
       my_xid tmp;
